refactor: split desktop & mobile
