+++ flutter/lib/mobile/pages/home_page.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_hbb/pages/chat_page.dart';
-import 'package:flutter_hbb/pages/server_page.dart';
-import 'package:flutter_hbb/pages/settings_page.dart';
-import '../common.dart';
+import 'package:flutter_hbb/mobile/pages/chat_page.dart';
+import 'package:flutter_hbb/mobile/pages/server_page.dart';
+import 'package:flutter_hbb/mobile/pages/settings_page.dart';
+import '../../common.dart';
 import '../widgets/overlay.dart';
 import 'connection_page.dart';
