+++ flutter/lib/mobile/widgets/gesture_help.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_hbb/common.dart';
 import 'package:toggle_switch/toggle_switch.dart';
 
-import '../models/model.dart';
+import '../../models/model.dart';
 
 class GestureIcons {
   static const String _family = 'gestureicons';
